separate migration for visits table
