+++ package.json
@@ -16,7 +16,7 @@
 		"prepare": "husky",
 		"db:shell": "sqlite3 database/local.db",
 		"db:kill": "rm -rf database/local.db",
-		"db:migrate": "tsx scripts/migrate.ts",
+		"db:migrate": "tsx scripts/migrate.ts && tsx database_visits/migrate.ts",
 		"db:seed": "tsx scripts/seed.ts",
 		"db:deduce": "tsx scripts/deduce.ts",
 		"db:test": "tsx scripts/test.ts",
